Implementing dynamic timestamp for the most recent status update.
Oh dear! It appears we are now employing a system whereby the timestamp of the latest published missive shall update dynamically. A sensible optimization, I must say, as it focuses computational resources where they are most likely to be of immediate interest to the user. One would hope this contributes to a more... shall we say... *fluid* experience for those viewing your pronouncements.